pyproject.toml
[build-system]
requires = ["setuptools>=66.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = { attr = "clever_faq.__version__" }

[project]
name = "clever_faq"
dynamic = ["version"]
description = "Test task for ray.app"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Danil Kovalev", email = "dan.kovalev2013@gmail.com" },
]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
dependencies = [
    "aioboto3==15.5.0",
    "alembic==1.17.2",
    "asgi-monitor==0.6.0",
    "bazario==0.4.2",
    "dishka==1.7.2",
    "fastapi==0.121.2",
    "langchain==1.0.5",
    "langchain-chroma==1.0.0",
    "langchain-experimental==0.4.0",
    "langchain-openai==1.0.3",
    "langsmith==0.4.42",
    "odfpy==1.4.1",
    "orjson==3.11.4",
    "pdfplumber==0.11.8",
    "python-docx==1.2.0",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "redis==6.4.0",
    "taskiq==0.11.20",
    "taskiq-aio-pika==0.4.4",
    "taskiq-redis==1.1.2",
    "uvicorn==0.38.0",
    "uvloop==0.21.0; sys_platform != 'win32'",
]

[project.optional-dependencies]
test = [
    "coverage[toml]==7.10.7",
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "httpx==0.28.1",
    "testcontainers[postgres,redis]==4.10.0"
]
lint = [
    "ruff==0.14.0",
    "bandit==1.8.6",
    "codespell==2.4.1",
    "mypy==1.18.2",
    "basedpyright==1.29.2",
    "semgrep==1.139.0"
]
dev = [
    "clever_faq[test,lint]",
    "pre-commit==4.3.0",
    "detect-secrets==1.5.0",
    "python-dotenv>=1.1.0",
    "pytest-env==1.2.0",
]

[project.urls]
Homepage = "https://github.com/C3EQUALZz/clever-faq"
Documentation = "https://github.com/C3EQUALZz/clever-faq/README.md"
Repository = "https://github.com/C3EQUALZz/clever-faq"
Issues = "https://github.com/C3EQUALZz/clever-faq/issues"

[tool.pytest.ini_options]
xfail_strict = true
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
markers = ["value_objects"]

[tool.basedpyright]
include = ["src", "tests"]
ignore = ["**/__pycache__", "**/alembic"]
pythonVersion = "3.12"
reportAny = false
reportUnusedCallResult = false

[tool.mypy]
mypy_path = "src"
plugins = "pydantic.mypy"
files = ["src"]
exclude = ["src/clever_faq/setup/config/"]
python_version = "3.12"

strict = true
pretty = true
warn_return_any = true
warn_unreachable = true
warn_unused_ignores = true
warn_redundant_casts = true
allow_redefinition = true
allow_untyped_calls = true
disallow_any_explicit = false
disallow_any_generics = false
disallow_subclassing_any = false
disallow_untyped_decorators = false
show_error_codes = true
show_error_context = true
show_column_numbers = true
check_untyped_defs = true
namespace_packages = true
ignore_missing_imports = true


[tool.ruff]
src = ["src"]
include = ["src/**.py", "tests/**.py"]
exclude = ["src/clever_faq/setup/config/**.py"]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "CPY001", "A005", "TRY003", "COM812", "FBT001", "FBT003", "PLR0913", "RUF001"]

[tool.ruff.lint.per-file-ignores]
"src/clever_faq/infrastructure/persistence/migrations/**" = ["ERA001", "INP001", "E501", "PLC0415"]
"src/clever_faq/presentation/http/v1/routes/documents/create_documents/handlers.py" = ["S311"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.isort]
known-first-party = ["src", "tests"]
no-lines-before = ["local-folder"]

[tool.coverage.run]
parallel = true
branch = true
concurrency = ["thread", "multiprocessing"]
source = ["src", "tests"]
omit = ["**/__init__.py"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_also = [
    'if TYPE_CHECKING:',
    "if __name__ == .__main__.:",
    '@(abc\.)?abstractmethod',
    "raise NotImplementedError",
    "pass",
    '\.\.\.',
]
omit = [
    'src/clever_faq/infrastructure/persistence/adapters/*.py',
    'src/clever_faq/infrastructure/persistence/alembic/env.py',
    '*/__init__.py',
]

[tool.semgrep]
paths = ["src"]
include = ["src"]

[tool.bandit]
targets = "src"

[tool.codespell]
skip = "*.pyi,*.pyc,./site,*.pdf"
